backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.13
python-dotenv==1.0.0
tavily-python==0.3.3
jinja2==3.1.4